Move module wrapper impl
